fixing warning and debug build refs #12, #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@18868 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/devices/MSDevice_Vehroutes.cpp
@@ -301,7 +301,7 @@ MSDevice_Vehroutes::generateOutput() const {
 
 const MSRoute*
 MSDevice_Vehroutes::getRoute(int index) const {
-    if (index < myReplacedRoutes.size()) {
+    if (index < (int)myReplacedRoutes.size()) {
         return myReplacedRoutes[index].route;
     } else {
         return 0;
